[Feat] UI & Layout Progress
The theme toggler will display as a button group on bigger screens.
The home page has a card with some basic bio info. There are ideas
of a layout.
Adjusted README, some html metadata info, and still don't really
have a need for a second page, but hopefully I can find a use for
Tanstack Router.