[Feat] MDX Ready
Configured the `/notes` index and `/notes/$noteId` dynamic
paths. The index gathers a list of each `.mdx` file's frontmatter
data to be displayed. The dynamic route will show the contents
of the approprite file. The frontmatter is read using a combination
of the `remark-frontmatter` and `remark-mdx-frontmatter` plugins.
The `remarkGfm` plugin has also been added to support certain MDX
syntax. There are other plugins I may look into in the future, such
as a code highlighting one. The styling is currently only handled by
the `@tailwindcss/typography` plugin (located in the index.css). I can
also make use of the `MdxProvider` from `@mdx-js/react` to further
customize the styling if need be.
Added `@tanstack/react-query` to fetch the post files in order to avoid
`useState` and `useEffect` and to take advantage of its cool features.
I could technically replace the `useQuery` from the `/notes` index route
with the tanstack router's built in loader functionality, but it would
require it to change to a non-lazy loaded file and lose its code splitting,
kind of that is. There's a way to have the best of both worlds.
I am now ready to start gathering all of the notes scattered around
various repos. Will need to improve the styling of these new pages.